Se auto calcula el pago automático cerrando un evento

+++ back/fisibet-pagos-abonos/src/main/java/com/microservicio/fisibet/infraestructure/port/spring/WalletSpringPort.java
@@ -3,5 +3,6 @@ package com.microservicio.fisibet.infraestructure.port.spring;
 import com.microservicio.fisibet.infraestructure.model.WalletModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface WalletSpringPort extends JpaRepository<WalletModel, Long> {
+public interface WalletSpringPort extends JpaRepository<WalletModel, Integer> {
+
 }
